+++ hrk_nasa_apis/lib/src/services/ssd_cneos/data/signature.dart
@@ -13,5 +13,5 @@ class Signature with _$Signature {
   factory Signature.fromJson(Map<String, dynamic> json) =>
       _$SignatureFromJson(json);
 
-  static const Map<String, String> fieldMap = _$$_SignatureFieldMap;
+  static const Map<String, String> fieldMap = _$$SignatureImplFieldMap;
 }