+++ hrk_nasa_apis/lib/src/services/ssd_cneos/data/signature.g.dart
@@ -1,7 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-// ignore_for_file: non_constant_identifier_names, constant_identifier_names
-
 // coverage:ignore-file
 
 part of 'signature.dart';
@@ -10,17 +8,18 @@ part of 'signature.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_Signature _$$_SignatureFromJson(Map<String, dynamic> json) => _$_Signature(
+_$SignatureImpl _$$SignatureImplFromJson(Map<String, dynamic> json) =>
+    _$SignatureImpl(
       version: json['version'] as String,
       source: json['source'] as String,
     );
 
-const _$$_SignatureFieldMap = <String, String>{
+const _$$SignatureImplFieldMap = <String, String>{
   'version': 'version',
   'source': 'source',
 };
 
-Map<String, dynamic> _$$_SignatureToJson(_$_Signature instance) =>
+Map<String, dynamic> _$$SignatureImplToJson(_$SignatureImpl instance) =>
     <String, dynamic>{
       'version': instance.version,
       'source': instance.source,