+++ hrk_nasa_apis/lib/src/services/ssd_cneos/sbdb_cad_api/data/query_parameters/sbdb_cad_query_parameters.dart
@@ -54,7 +54,7 @@ class SbdbCadQueryParameters with _$SbdbCadQueryParameters {
       _$SbdbCadQueryParametersFromJson(json);
 
   static const Map<String, String> fieldMap =
-      _$$_SbdbCadQueryParametersFieldMap;
+      _$$SbdbCadQueryParametersImplFieldMap;
 
   static DateTime get dateMinDefault => HrkDateTime.today();
   static DateTime get dateMaxDefault =>